1. Added the functionality for closing the Ride and getting the distance. 2. Added the idling resource such that we do not wait for long time for the SMS.

+++ app/src/androidTest/java/com/vave/getbike/utils/GetBikeTestUtils.java
@@ -21,7 +21,7 @@ public class GetBikeTestUtils {
                 TextView textView = (TextView) item;
                 boolean result = false;
                 if (textView.getText() != null) {
-                    if (Integer.parseInt(textView.getText() + "") > 0) {
+                    if (Double.parseDouble(textView.getText() + "") > 0) {
                         result = true;
                     }
                 }
